backend/pyproject.toml
[tool.poetry]
name = "mesabite"
version = "0.1.0"
description = ""
authors = ["Zoltan Kohalmy <zoltankohalmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
Pillow = "^10.1.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.6.0"
pytest-sugar = "^0.9.7"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
